Refactor pkg/kubectl/conditions.go to remove file
